backend/requirements.txt
flask
flask-cors
pandas
pymongo
pinecone
sentence-transformers
python-dotenv
nltk
scikit-learn
joblib
google-generativeai
gunicorn